Centralize shell step execution
Keep workflow and composite shell evaluation, working-directory handling, container translation, and process launch behind one runtime boundary without changing their existing semantics.

Amp-Thread-ID: https://ampcode.com/threads/T-01a06089-3908-723b-946f-73dcda7f0957
